SLRMss-package {SLRMss}	R Documentation

Symmetric Linear Regression Models for small samples

Description

A function to obtain ordinary and modified statistics for symmetrical linear regression models with small samples. The supported ordinary statistics include Wald,score, likelihood ratio and gradient and the modified statistics include score, likelihood ratio and gradient.

Details

The DESCRIPTION file:

Package: 	 SLRMss

Type: 	 Package

Title: 	 Symmetric Linear Regression Models for small samples

Version: 	 1.0

Date: 	 2020-01-25

Author: 	 Ivonaldo Silvestre da Silva-Junior [aut,cre], Francisco Moises C. de Medeiros [aut]

Maintainer: 	 Ivonaldo Silvestre da Silva-Junior <ivosilvestresjr@gmail.com>

Description: 	 A function to obtain ordinary and modified statistics for symmetrical linear regression models with small samples. The supported ordinary statistics include Wald,score, likelihood ratio and gradient and the modified statistics include score, likelihood ratio and gradient.

License: 	 What license is it under?

NeedsCompilation: 	 no

Packaged: 	 2020-05-08 18:23:38 UTC; ivosi

Depends: 	 R (>= 3.5.0), ssym, nleqslv

Index of help topics:


SLRMss-package          Symmetric Linear Regression Models for small
                        samples
cheese                  Cheese Data set
orange                  Orange Data Set
plot.SLRMss             plot.SLRMss
print.SLRMss            print.SLRMss
slrmss                  Symmetric Linear Regression Models for Small
                        Samples
summary.SLRMss          summary.SLRMss

Author(s)

Ivonaldo Silvestre da Silva-Junior [aut,cre], Francisco Moises C. de Medeiros [aut]

Maintainer: Ivonaldo Silvestre da Silva-Junior <ivosilvestresjr@gmail.com>

References

Medeiros, F. M. C and Ferrari, S. L. P. (2017 Small-sample testing inference in symmetric and log-symmetric linear regression models, Statistica Neerlandica 2017.

Examples


library(SLRMss)
data(orange)
fit=slmrss(emulsion~arabicgum+xanthangum+orangeoil,
family="Student",xi=3,testingbeta="xanthangum",data=orange)
summary(fit)

[Package SLRMss version 1.0 Index]