cheese {SLRMss}	R Documentation

Cheese Data set

Description

The Nateghi's cheese data set gives measurements of the cheddar cheese textures.

Usage

data("cheese")

Format

A data frame with 16 observations on the following 4 variables.

cohe

the cohesiveness

fat

the fat percentage

xangum

the xanthan gum percentage

sodcase

the sodium caseinate percentage

Source

Nateghi, L., Roohinejad, S., Totosaus, A., Mirhosseini, H., Shuhaimi, M., Meimandipour, A., Omidizadeh, A., Manap, M.Y.A (2012). Optimization of textural properties and formulation of reduced fat Cheddar cheeses containing fat replacers. Journal of Food, Agriculture & Environment.

Examples


data(cheese)
plot(cheese)
plot(exp(cheese$cohe),cheese$fat)
fit=slrmss(exp(cohe)~fat+xangum,family="Normal",testingbeta="xangum",statistic="Gradient",data=cheese)
summary(fit)

[Package SLRMss version 1.0 Index]