orange {SLRMss}	R Documentation

Orange Data Set

Description

The Mirhosseini's e Tan's data set gives the measurements of the physicochemical properties of emulsion in orange drinks.

Usage

data("orange")

Format

A data frame with 20 observations on the following 6 variables.

emulsion

the emulsion density in g/cm^3

arabicgum

the arabic gum quantity in g

xanthangum

the xanthan gum quantity in g

orangeoil

the orange oil quantity in g

Source

Mirhosseini, H., Tan, C.P. (2010). Discrimination of orange beverage emulsions with different formulations using multivariate analysis. Journal of the Science of Food and Agriculture.

Examples


data(orange)
plot(orange)
fit=slrmss(emulsion~arabicgum+xanthangum+orangeoil,family="Student",xi=3,testingbeta="xanthangum",data=orange)
summary(fit)
plot(fit)

[Package SLRMss version 1.0 Index]