slrmss {SLRMss}	R Documentation

Symmetric Linear Regression Models for Small Samples

Description

Computes Wald, Likelihood-ratio, Score or Gradient of symmetrical linear regression models. Also computes modified Likelihood-ratio, Score and Gradient for small samples.

Usage


slrmss(formula, family, xi, statistic, testingbeta, data=NULL)

Arguments

formula
an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted.

family
a description of the error distribution to be used in the model. There are four supported families, Normal, t-Student, Power Exponential and Logistic ("Normal", "Student", "Powerexp" and "Logistic", respectively)

xi
an extra parameter of some specified error distribution. For t-Student is a positive value and for Power Exponential is a real number between -1 and 1/3.

statistic
the statistic which will be used. It includes "Wald", "LR", "Score" or "Gradient".

testingbeta
an vector containing the names of the variables to be testing.

data
an optional data frame containing the variables in the model.

Value

beta.coefficients
a matrix with the estimated position parameters under alternative hypothesis.

phi
a numeric value with the estimated precision paramater under alternative hypothesis.

beta.coefficients.h0
a matrix with the estimated position parameters under null hypothesis.

phi.h0
a numeric value with the estimated precision paramater under null hypothesis.

beta.fitted
a vector with the fitted values of the model.

null.hypothesis
the description of the null hypothesis.

statistics
a matrix with the selected statistics and theirs p-values.

statistic.distribution
the name of the statistics' distribution used to test null hypothesis. It always return "Chi-Squared".

df
the degrees of freedom of the statistics' distribution. It's the length of the testingbeta vector.

std.residuals
the standardized residuals.

AICc
the corrected Akaike Information Criterion for small samples.

BIC
the Bayesian Information Criterion.

Author(s)

Ivonaldo Silvestre da Silva-Junior <ivonaldojunior@hotmail.com.br> and Francisco Moises C. de Medeiros

References

Medeiros, F. M. C and Ferrari, S. L. P. (2017 Small-sample testing inference in symmetric and log-symmetric linear regression models, Statistica Neerlandica 2017.

Examples


#Using the orange data frame to fit the emulsion density and testing the xanthan gum quantity.
#
#data(orange)
#
#slrmss(emulsion~arabicgum+xanthangum,family="Normal",testingbeta="xanthangum",data=orange)
#
#slrmss(emulsion~arabicgum+xanthangum+orangeoil,family="Normal",testingbeta="xanthangum",data=orange)
#
#fit<-slrmss(emulsion~arabicgum+xanthangum+orangeoil,family="Student",xi=3,testingbeta="xanthangum",data=orange)
#summary(fit)
#plot(fit)
###############################################################
###############################################################
#Using the cheese data frame to fit the cohesiveness and test xanthan gum percentage and sodium caseinate percentage.
#
#data(cheese)
#
#fit<-slrmss(exp(cohe)~fat+xangum+sodcase,family="Normal",testingbeta=c("xangum","sodcase"),statistic="Gradient",data=cheese)
#summary(fit)
#plot(fit)



[Package SLRMss version 1.0 Index]