summary.SLRMss {SLRMss}	R Documentation

summary.SLRMss

Description

It displays the summary of the fitted model. It includes parameter estimates under both, null and alternative hypothesis, corrected Akaike information criterion, bayesian information criterion and choosed statistics.

[Package SLRMss version 1.0 Index]